# docs/killrvideo_openapi.yaml
openapi: 3.1.0
info:
  title: KillrVideo v2 - Monolith Backend
  version: 2.0.0-alpha-local
paths:
...
  /api/v1/videos/{video_id_path}/comments:
    post:
      tags:
      - Comments & Ratings
      summary: Add comment to video
      description: Endpoint for viewers to add a comment to a READY video.
      operationId: post_comment_to_video_api_v1_videos__video_id_path__comments_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Comments & Ratings
      summary: List comments for video
      operationId: list_comments_video_api_v1_videos__video_id_path__comments_get
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 10
          title: Pagesize
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/{user_id_path}/comments:
    get:
      tags:
      - Comments & Ratings
      summary: List comments by user
      operationId: list_comments_user_api_v1_users__user_id_path__comments_get
      parameters:
      - name: user_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id Path
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 10
          title: Pagesize
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/videos/{video_id_path}/ratings:
    post:
      tags:
      - Comments & Ratings
      summary: Rate a video (create or update)
      description: Upsert a rating (1-5) for the specified video by the current viewer.
      operationId: post_rating_video_api_v1_videos__video_id_path__ratings_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingCreateOrUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Comments & Ratings
      summary: Get aggregate rating plus current user's rating (optional)
      operationId: get_rating_summary_video_api_v1_videos__video_id_path__ratings_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateRatingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AggregateRatingResponse:
      properties:
        videoId:
          type: string
          format: uuid
          title: Videoid
        averageRating:
          anyOf:
          - type: number
          - type: 'null'
          title: Averagerating
        totalRatingsCount:
          type: integer
          title: Totalratingscount
          default: 0
        currentUserRating:
          anyOf:
          - type: integer
          - type: 'null'
          title: Currentuserrating
      type: object
      required:
      - videoId
      title: AggregateRatingResponse
    CommentCreateRequest:
      properties:
        text:
          type: string
          maxLength: 1000
          minLength: 1
          title: Text
      type: object
      required:
      - text
      title: CommentCreateRequest
      description: Payload for creating a new comment.
    CommentResponse:
      properties:
        commentid:
          type: string
          format: uuid
          title: Commentid
        videoid:
          type: string
          format: uuid
          title: Videoid
        userid:
          type: string
          format: uuid
          title: Userid
        comment:
          type: string
          title: Comment
        sentiment_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Sentiment Score
        firstName:
          anyOf:
          - type: string
          - type: 'null'
          title: Firstname
        lastName:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastname
      type: object
      required:
      - commentid
      - videoid
      - userid
      - comment
      title: CommentResponse
      description: API response representation for a single comment.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PaginatedResponse:
      properties:
        data:
          items: {}
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: PaginatedResponse
...
    Pagination:
      properties:
        currentPage:
          type: integer
          title: Currentpage
        pageSize:
          type: integer
          title: Pagesize
        totalItems:
          type: integer
          title: Totalitems
        totalPages:
          type: integer
          title: Totalpages
      type: object
      required:
      - currentPage
      - pageSize
      - totalItems
      - totalPages
      title: Pagination
    RatingCreateOrUpdateRequest:
      properties:
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
          description: Rating value 1-5
      type: object
      required:
      - rating
      title: RatingCreateOrUpdateRequest
    RatingResponse:
      properties:
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
          description: Rating value 1-5
        videoid:
          type: string
          format: uuid
          title: Videoid
        userid:
          type: string
          format: uuid
          title: Userid
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - rating
      - videoid
      - userid
      title: RatingResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/users/login